+++ src/Resume.tsx
@@ -3,7 +3,7 @@ import { Education } from "./components/education/education";
 import { Employment } from "./components/Employment";
 import { Interests } from "./components/Interests";
 import { PersonalDetails } from "./components/personal-details/personal-details";
-import { Qualifications } from "./components/Qualifications";
+import { Qualifications } from "./components/qualifications/qualifications";
 import { References } from "./components/References";
 import { Skills } from "./components/Skills";
